SPL: Use token-2022-interface crate (#7298)
#### Problem

As outlined in #7256, we need to get Agave off of crates that depend
back on Agave.

spl-token-2022 is one of those crates

#### Summary of changes

Use spl-token-2022-interface instead. To go with this, we had to also
upgrade to confidential-transfer-proof-extract v0.4.0, which requires
modifications to some tests.

Everything should be very straightforward otherwise, and we get to see
some red lines in the lockfile!